Manage security rules by code to avoid them being cleared out after a deplyoment
